+++ frost-core/src/frost/keys/dkg.rs
@@ -51,8 +51,9 @@ pub mod round1 {
 
     /// The package that must be broadcast by each participant to all other participants
     /// between the first and second parts of the DKG protocol (round 1).
-    #[derive(Clone)]
+    #[derive(Clone, PartialEq, Eq)]
     #[cfg_attr(feature = "serde", derive(serde::Serialize, serde::Deserialize))]
+    #[cfg_attr(feature = "serde", serde(deny_unknown_fields))]
     pub struct Package<C: Ciphersuite> {
         /// The identifier of the participant who is sending the package (i).
         pub sender_identifier: Identifier<C>,
@@ -93,8 +94,9 @@ pub mod round2 {
     /// # Security
     ///
     /// The package must be sent on an *confidential* and *authenticated* channel.
-    #[derive(Clone)]
+    #[derive(Clone, PartialEq, Eq)]
     #[cfg_attr(feature = "serde", derive(serde::Serialize, serde::Deserialize))]
+    #[cfg_attr(feature = "serde", serde(deny_unknown_fields))]
     pub struct Package<C: Ciphersuite> {
         /// The identifier of the participant that generated the package (i).
         pub sender_identifier: Identifier<C>,